created separate files for genAgg
